Lesson 22: Added Y-Sort ordering for all BaseMapLayer Nodes and children
Added Y sort ordering for all the sprites in the BaseMapLayer, so they render correctly and not on top of each other.   To do so we created a new YSortOrder Node in Main, inserted all the Map Layers as its children and enabled Ysort on the whole chain.  To make it so that the sand/terrain tiles don't overlap with the building sprites, we changed the building sprite y position, so it is above the one of the sand tile on a given cell