Lesson 25: added counter method for collected reources
We added a method that counts how many wood resources are gathered by houses in range. // We implemented a method to suppress double dipping of a single resource. A custom signal is then given whenever a new resource is gathered on which we can listen from other nodes to update the game node status