Lesson 26: added a simple UI to handle buttons
introduction to UI Nodes
we created a new UI node that can handle all the buttons, which will be expanded on later. This declutters the Main scene and allows for a node which handles all the UI features.
For the UI we used a container node, which can be usde to contain all the different buttons and allows for a smoother and simpler editing of the button location.
To also declutter the main C# script we also created a new UI script which will handle all UI-Related EventHandling (i.e. the Eventhandlers)