Lesson 29: Added building placement ghost
We added a building placement ghost that keep track of the building selected based on button pressed and displays a sprite of it on the hovered tile.
To do so we created a separate sprite2D instance under each building scene, which specifically contains the sprite to be displayed and the necessary offset position of it. This allows for us to pass this one dynamically to the ghostscene as well and always having the right positioning on the tile displayed