Lesson 44: Buildings can now occupy multiple tiles
we exportet a new building resource called Dimensions to assign the dimensions of hte building once placed on the grid.
We then assigned a two by two area to our buildings.
Code adjustments so different methods don't check for tiles but for the building are instead. To do so we changed the tileset to a areaset and the holding list type from a Vector2I to a Rect2I. Latter holds vector data to the rectangle positions, i.e upper left and lower right corner of the rectangle