Lesson 47: buildings now only on single elevation layer
Buildings could be placed over different elevation layers. To fix this uwanted behaviour, we divided the scenes into multiple elevation layers by introducin new nodes for every single elevation layer. Under each eleavtion node we input all the tile and resource Layers.
These elevation nodes can the be referenced in the code and using a touple we assign all the different tileMapLayers to a differetn elevationLayer. This touple we can then use when checking for buildable tiles to ensure that a building is only placed on tiles of the same eleavtion layer.
In the tileAreaBuildable method we now get the to-be-placed Building's area and check whether the tiles are buildable, all the tiles in the building area coincide with the to be placed area and WETHER THE BUILDING AREA IS ALL ON THE SAME LEVEL as of the hovered/ selected position