Lesson 48: Added an extension Method to the GoDot Rect2I class
Extension methods are a way that can add methods to classes that are not changeble, like for example GoDot classes. This does not actually add a new method, but for all means it works as one and can be called, when using e.g. the Rect2I class as one of its normal methods.
This we did to a part of the code that we often apply in conjunction to the Rect2I class, where we return all the tilePositions within the passed Rect2I